requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
